Safari-611.1.10/LayoutTests/fast/events/touch/ios/touch-event-regions/active-passive-nesting-expected.txt

Passive
Active
Active
Passive
touchstart
162, 74 - 270, 182
48, 258 - 226, 436
touchstart

touchend

touchforcechange

passive
48, 40 - 226, 218
162, 292 - 270, 400
